requirements.txt
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.9.1
cffi==2.0.0
click==8.1.8
colorama==0.4.6
cryptography==39.0.1
django==4.2.24
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
dnspython==2.7.0
email-validator==2.3.0
et-xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.104.1
h11==0.16.0
hdbcli==2.21.26
httptools==0.6.4
idna==3.10
numpy==2.0.2
openpyxl==3.1.5
pandas==2.3.2
psycopg2==2.9.10
psycopg2-binary==2.9.9
pycparser==2.23
pydantic==2.5.0
pydantic-core==2.14.1
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
pyyaml==6.0.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.27.0
typing-extensions==4.15.0
tzdata==2025.2
uvicorn==0.24.0
watchfiles==1.1.0
websockets==15.0.1
